+++ contrib/platform/src/com/sun/jna/platform/win32/User32.java
@@ -2445,4 +2445,44 @@ public interface User32 extends StdCallLibrary, WinUser, WinNT {
      * error information, call GetLastError.</p>
      */
     public int RegisterClipboardFormat(String formatName);
+    
+    /**
+     * Retrieves the window handle to the active window attached to the 
+     * calling thread's message queue.
+     * 
+     * @return Type: HWND The return value is the handle to the active 
+     * window attached to the calling thread's message queue. Otherwise, 
+     * the return value is NULL.
+     */
+    public HWND GetActiveWindow();
+  
+	/**
+	 * Sends the specified message to a window or windows. 
+	 * The SendMessage function calls the window procedure for the specified window and 
+	 * does not return until the window procedure has processed the message.
+	 * 
+ 	 * To send a message and return immediately, use the SendMessageCallback or SendNotifyMessage function. 
+ 	 * To post a message to a thread's message queue and return immediately, 
+ 	 * use the PostMessage or PostThreadMessage function.
+	 * 
+	 * @param hWnd A handle to the window whose window procedure will receive the message. 
+	 * 		If this parameter is HWND_BROADCAST ((HWND)0xffff), the message is sent to all 
+	 * 		top-level windows in the system, including disabled or invisible unowned windows, 
+	 * 		overlapped windows, and pop-up windows; but the message is not sent to child windows.
+	 * 		Message sending is subject to UIPI. The thread of a process can send messages 
+	 * 		only to message queues of threads in processes of lesser or equal integrity level.
+	 * 
+	 * @param msg The message to be sent.
+	 * 		For lists of the system-provided messages, see System-Defined Messages.
+	 * @param wParam Additional message-specific information.
+	 * @param lParam Additional message-specific information.
+	 * 
+	 * @return The return value specifies the result of the message processing; it depends on the message sent.
+	 * 
+	 * Two classic usage : 
+	 *  - with a WM_USER+x msg value : wParam and lParam are numeric values
+	 *  - with a WM_COPYDATA msg value : wParam is the length of the structure and lParam a pointer to a COPYDATASTRUCT 
+	 */
+    LRESULT SendMessage(HWND hWnd, int msg, WPARAM wParam, LPARAM lParam);
+	
 }